requirements.txt
requests==2.31.0
python-dotenv==1.0.0
playwright==1.44.0
schedule==1.2.1
gspread==6.1.2
google-auth==2.29.0
urllib3==2.0.7